gcp/process_mails/requirements.txt
google-api-python-client 
google-auth-httplib2 
google-auth-oauthlib
functions-framework==3.*
pymongo
pytz
requests
sentry-sdk
tiktoken
openai
pinecone-client
langchain
langchain-core
langchain-pinecone
langchain-openai
langchain-community
langgraph>0.2.27
pydantic
six
mistralai
chromadb
beautifulsoup4
icalendar